use memq LCM provider
As described in an updated version of the API documentation (e.g.,
[v1.4.0]), the memory queue test provider:

> is primarily useful for testing, especially unit testing code
> that uses LCM.  It provides a pub/sub interface that is private to the LCM
> instance, and does _not_ provide any interprocess communication, or even
> inter-instance communication.  Use this provider to implement unit tests
> that require deterministic and predictable behavior that is independent of
> a system's network configuration.

So unless we share an instance between threads, the threads will not
communicate via LCM.

I was hoping to use this as a shortcut for simulating
faster-than-realtime systems both between threads and between processes,
but it looks like that traffic will still have to go over the network
stack, and I'll just use a different `udpm` provider. For some projects,
that might mean revising code to rely on the environment variable
`LCM_DEFAULT_URL` rather than a provider that is hard-coded or
`#define`d in a header, and it could complicate split-ttl architectures.
Honestly most people using LCM won't care about split-ttl, so I
understand needing to jump through a few extra hoops if that's what the
project needs.